test(gateway): share streaming root-work test helpers
createPostHandlerGate and waitForRootWorkCount were duplicated verbatim in both
streaming HTTP test files.
